move columnchecker to the common package
