+++ vertx-pg-client/src/test/java/io/vertx/pgclient/data/NullSimpleCodecTest.java
@@ -1,6 +1,7 @@
 package io.vertx.pgclient.data;
 
 import io.vertx.pgclient.PgConnection;
+import io.vertx.sqlclient.ColumnChecker;
 import io.vertx.sqlclient.Row;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
