+++ vertx-pg-client/src/test/java/io/vertx/pgclient/data/TsTypesSimpleCodecTest.java
@@ -3,6 +3,7 @@ package io.vertx.pgclient.data;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.pgclient.PgConnection;
+import io.vertx.sqlclient.ColumnChecker;
 import io.vertx.sqlclient.Row;
 import io.vertx.sqlclient.Tuple;
 import org.junit.Test;
